+++ pwned-server/loadtest/loadtest.cpp
@@ -96,7 +96,7 @@ void progress(const boost::system::error_code&, boost::asio::steady_timer *t, Ht
 int main(int argc, const char *argv[])
 {
   static const std::string DefaultAddress = "http://127.0.0.1:31337/v1/pwned/api/lookup";
-  static const int DefaultNumThreads = std::thread::hardware_concurrency();
+  static const int DefaultNumThreads = int(std::thread::hardware_concurrency());
   static const int DefaultNumWorkers = 64;
   static const int DefaultRuntimeSecs = 10;
   std::string inputFilename;
@@ -183,7 +183,7 @@ int main(int argc, const char *argv[])
     timer.async_wait(boost::bind(progress, boost::asio::placeholders::error, &timer, &workers.front(), double(runtimeSecs)));
 
     std::vector<std::thread> threads;
-    threads.reserve(numThreads);
+    threads.reserve(size_t(numThreads));
     for (auto i = 0; i < numThreads; ++i)
     {
       threads.emplace_back(
